Replaceable jl_parse as single parser entry point
`jl_parse` calls a C function pointer which can be replaced with
jl_set_parser(). This should make it possible to replace the flisp
parser with julia code. Currently `jl_parse` takes Strings rather than C
buffers to make it more natural to implement in Julia code.

TODO: Could revisit this to decide whether it's desirable to require a
`String` or whether we should have non-copying `unsafe_wrap` and
`IOBuffer` for parsing.